fix: moved map into quick-map component
