+++ src/app/pages/mapsTool/mapsTool.component.html
@@ -1,4 +1,3 @@
 <h1>maps-tool component</h1>
-<a [routerLink]="ROUTES.QUICK_MAPS.route()">BMGF MAPS</a>
-<div id="map" style="height: 300px;">
-</div>
+<a [routerLink]="ROUTES.QUICK_MAPS.route()">QUICK MAPS</a>
+
