+++ src/app/pages/mapsTool/mapsTool.component.ts
@@ -1,5 +1,4 @@
-import { Component, OnInit } from '@angular/core';
-import * as L from 'leaflet';
+import { Component } from '@angular/core';
 import { AppRoutes } from 'src/app/app-routes';
 
 @Component({
@@ -7,13 +6,8 @@ import { AppRoutes } from 'src/app/app-routes';
   templateUrl: './mapsTool.component.html',
   styleUrls: ['./mapsTool.component.scss'],
 })
-export class MapsToolComponent implements OnInit {
+export class MapsToolComponent {
   public ROUTES = AppRoutes;
 
-  ngOnInit(): void {
-    const myMap = L.map('map').setView([51.505, -0.09], 13);
-    L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {
-      attribution: '© <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
-    }).addTo(myMap);
-  }
+
 }
