+++ src/app/pages/quickMaps/quickMaps.component.html
@@ -1,1 +1,4 @@
 <h1>quick-maps component</h1>
+
+<div id="map" style="height: 300px;">
+</div>
